working on current project in transactions controller

+++ src/main/java/info/podkowinski/sandra/financescanner/transaction/TransactionController.java
@@ -24,7 +24,6 @@ import java.time.LocalDate;
 import java.util.*;
 
 
-
 @Controller
 @RequestMapping("/transaction")
 public class TransactionController {
@@ -45,15 +44,14 @@ public class TransactionController {
     }
 
     @RequestMapping("/list")
-    public String transaction(Model model
-//                              @AuthenticationPrincipal CurrentUser customUser
-    ) {
-        Project project2 = projectService.findById(2l);
-        List<Transaction> transactionsList = transactionService.findByProjectId(2l);
-        List<Account> accountsList = accountService.findByProjectId(2l);
-        List<Category> categoriesList = categoryService.findByProjectId(2l);
-        List<Integer> years = transactionService.findYearsByProjectId(2l);
-        HashMap<Long, List<String>> transactionCategory = transactionService.transactionIdCategories(2l);
+    public String transaction(Model model, @AuthenticationPrincipal CurrentUser currentUser) {
+
+        Project project = currentUser.getUser().getCurrentProject();
+        List<Transaction> transactionsList = transactionService.findByProjectId(project.getId());
+        List<Account> accountsList = accountService.findByProjectId(project.getId());
+        List<Category> categoriesList = categoryService.findByProjectId(project.getId());
+        List<Integer> years = transactionService.findYearsByProjectId(project.getId());
+        HashMap<Long, List<String>> transactionCategory = transactionService.transactionIdCategories(project.getId());
         model.addAttribute("tl", transactionsList);
         model.addAttribute("bl", accountsList);
         model.addAttribute("categoriesList", categoriesList);
@@ -63,19 +61,19 @@ public class TransactionController {
     }
 
     @GetMapping("/fileimport")
-    public String fileimport(Model model) {
-        Project project1 = projectService.findById(2l);
-        List<Account> accountsList = accountService.findByProjectId(2l);
-        List<CsvSettings> csvSettingsList = csvSettingsService.findByProjectId(project1.getId());
+    public String fileimport(Model model, @AuthenticationPrincipal CurrentUser currentUser) {
+        Project project = currentUser.getUser().getCurrentProject();
+        List<Account> accountsList = accountService.findByProjectId(project.getId());
+        List<CsvSettings> csvSettingsList = csvSettingsService.findByProjectId(project.getId());
         model.addAttribute("csvSettingsList", csvSettingsList);
         model.addAttribute("accountsList", accountsList);
         return "file-import";
     }
 
     @PostMapping("/fileimport")
-    public String fileimportPost(HttpServletRequest request) throws IOException, ServletException, ParseException, CsvValidationException {
+    public String fileimportPost(HttpServletRequest request, @AuthenticationPrincipal CurrentUser currentUser) throws IOException, ServletException, ParseException, CsvValidationException {
         Part filePart = request.getPart("fileToUpload");
-        Project project1 = projectService.findById(2l);
+        Project project = currentUser.getUser().getCurrentProject();
         int datePosition = Integer.parseInt(request.getParameter("datePosition")) - 1;
         int descriptionPosition = Integer.parseInt(request.getParameter("descriptionPosition")) - 1;
         int partyPosition = Integer.parseInt(request.getParameter("partyPosition")) - 1;
@@ -85,15 +83,15 @@ public class TransactionController {
         String importName = request.getParameter("importName");
         Long account = Long.parseLong(request.getParameter("selectAccount"));
         transactionService.scanDocument(filePart.getInputStream(), datePosition, descriptionPosition,
-                partyPosition, amountPosition, separator, skippedLines, importName, account, project1);
+                partyPosition, amountPosition, separator, skippedLines, importName, account, project);
         return "redirect:/transaction/list";
     }
 
     @GetMapping("/add")
-    public String add(Model model) {
-        Project project1 = projectService.findById(2l);
-        List<Category> categories = categoryService.findByProjectId(2l);
-        List<Account> accounts = accountService.findByProjectId(2l);
+    public String add(Model model, @AuthenticationPrincipal CurrentUser currentUser) {
+        Project project = currentUser.getUser().getCurrentProject();
+        List<Category> categories = categoryService.findByProjectId(project.getId());
+        List<Account> accounts = accountService.findByProjectId(project.getId());
         Transaction transaction = new Transaction();
         model.addAttribute("categories", categories);
         model.addAttribute("accounts", accounts);
@@ -103,26 +101,24 @@ public class TransactionController {
 
     //todo frontend validation
     @PostMapping("/add")
-    public String addPost(HttpServletRequest request, @ModelAttribute Transaction transaction1) {
-        Project project1 = projectService.findById(2l);
-        transaction1.setProject(project1);
+    public String addPost(HttpServletRequest request, @ModelAttribute Transaction transaction1, @AuthenticationPrincipal CurrentUser currentUser) {
+        Project project = currentUser.getUser().getCurrentProject();
+        transaction1.setProject(project);
         transaction1.setImportName("Imported manually on " + LocalDate.now());
-
         String date = request.getParameter("transactionDate");
         System.out.println(date);
         transaction1.setTransactionDate(LocalDate.parse(date));
         System.out.println(transaction1.transactionDate);
-
         transactionService.save(transaction1);
 
         return "redirect:/transaction/table/gettransaction/" + transaction1.id;
     }
 
     @GetMapping("/edit/{transactionId}")
-    public String edit(Model model, @PathVariable Long transactionId) {
-        Project project1 = projectService.findById(2l);
-        List<Category> categories = categoryService.findByProjectId(2l);
-        List<Account> accounts = accountService.findByProjectId(2l);
+    public String edit(Model model, @PathVariable Long transactionId, @AuthenticationPrincipal CurrentUser currentUser) {
+        Project project = currentUser.getUser().getCurrentProject();
+        List<Category> categories = categoryService.findByProjectId(project.getId());
+        List<Account> accounts = accountService.findByProjectId(project.getId());
         Transaction transaction = transactionService.findById(transactionId);
         model.addAttribute("categories", categories);
         model.addAttribute("accounts", accounts);
@@ -132,14 +128,12 @@ public class TransactionController {
 
     //todo frontend validation
     @PostMapping("/edit/{transactionId}")
-    public String editPost(HttpServletRequest request, @ModelAttribute Transaction transaction1) {
-        Project project1 = projectService.findById(2l);
-        transaction1.setProject(project1);
+    public String editPost(HttpServletRequest request, @ModelAttribute Transaction transaction1, @AuthenticationPrincipal CurrentUser currentUser) {
+        Project project = currentUser.getUser().getCurrentProject();
+        transaction1.setProject(project);
         transaction1.setImportName("Imported manually on " + LocalDate.now());
-
         String date = request.getParameter("transactionDate");
         transaction1.setTransactionDate(LocalDate.parse(date));
-
         System.out.println(transaction1.transactionDate);
         transactionService.save(transaction1);
         return "redirect:/transaction/table/gettransaction/" + transaction1.id;
@@ -148,25 +142,25 @@ public class TransactionController {
 
     @ResponseBody
     @GetMapping("/removeallcategories/{transactionId}")
-    public String removeAllCategories(@PathVariable Long transactionId) {
-        Project project1 = projectService.findById(2l);
+    public String removeAllCategories(@PathVariable Long transactionId, @AuthenticationPrincipal CurrentUser currentUser) {
+        Project project = currentUser.getUser().getCurrentProject();
         Transaction transaction = transactionService.findById(transactionId);
-        if (transaction.getProject() == project1) {
-                transaction.rejectedCategories.addAll(transaction.getPendingCategories());
-                transaction.rejectedCategories.addAll(transaction.getCategories());
-                transaction.setCategories(null);
-                transaction.setPendingCategories(null);
-                transactionService.save(transaction);
-            }
+        if (transaction.getProject() == project) {
+            transaction.rejectedCategories.addAll(transaction.getPendingCategories());
+            transaction.rejectedCategories.addAll(transaction.getCategories());
+            transaction.setCategories(null);
+            transaction.setPendingCategories(null);
+            transactionService.save(transaction);
+        }
         return "";
     }
 
     @ResponseBody
     @GetMapping("/acceptallsuggestions/{transactionId}")
-    public String acceptAllSuggestions(@PathVariable Long transactionId) {
-        Project project1 = projectService.findById(2l);
+    public String acceptAllSuggestions(@PathVariable Long transactionId, @AuthenticationPrincipal CurrentUser currentUser) {
+        Project project = currentUser.getUser().getCurrentProject();
         Transaction transaction = transactionService.findById(transactionId);
-        if (transaction.getProject() == project1) {
+        if (transaction.getProject() == project) {
             transaction.categories.addAll(transaction.getPendingCategories());
             transaction.setPendingCategories(null);
             transactionService.save(transaction);
@@ -176,10 +170,10 @@ public class TransactionController {
 
     @ResponseBody
     @GetMapping("/rejectallsuggestions/{transactionId}")
-    public String rejectAllSuggestions(@PathVariable Long transactionId) {
-        Project project1 = projectService.findById(2l);
+    public String rejectAllSuggestions(@PathVariable Long transactionId, @AuthenticationPrincipal CurrentUser currentUser) {
+        Project project = currentUser.getUser().getCurrentProject();
         Transaction transaction = transactionService.findById(transactionId);
-        if (transaction.getProject() == project1) {
+        if (transaction.getProject() == project) {
             transaction.rejectedCategories.addAll(transaction.getPendingCategories());
             transaction.setPendingCategories(null);
             transactionService.save(transaction);
@@ -189,10 +183,10 @@ public class TransactionController {
 
     @ResponseBody
     @GetMapping("/addcategory/{transactionId}/{categoryId}")
-    public String addCategory(@PathVariable Long transactionId, @PathVariable Long categoryId) {
-        Project project1 = projectService.findById(2l);
+    public String addCategory(@PathVariable Long transactionId, @PathVariable Long categoryId, @AuthenticationPrincipal CurrentUser currentUser) {
+        Project project = currentUser.getUser().getCurrentProject();
         Transaction transaction = transactionService.findById(transactionId);
-        if (transaction.getProject() == project1) {
+        if (transaction.getProject() == project) {
             transaction.addCategory(categoryService.findById(categoryId));
             transactionService.save(transaction);
         }
@@ -201,15 +195,13 @@ public class TransactionController {
 
     @ResponseBody
     @GetMapping("/removecategory/{transactionId}/{categoryId}")
-    public String removeCategory(@PathVariable Long transactionId, @PathVariable Long categoryId) {
-        Project project1 = projectService.findById(2l);
+    public String removeCategory(@PathVariable Long transactionId, @PathVariable Long categoryId, @AuthenticationPrincipal CurrentUser currentUser) {
+        Project project = projectService.findCurrentByUser(currentUser.getUser());
         Category categoryToRemove = categoryService.findById(categoryId);
         Transaction transaction = transactionService.findById(transactionId);
-        if (transaction.getProject() == project1) {
-            if(transaction.pendingCategories.contains(categoryToRemove)){
+        if (transaction.getProject() == project) {
+            if (transaction.pendingCategories.contains(categoryToRemove)) {
                 transaction.removeCategory(categoryToRemove);
-            } else {
-                transaction.removeCategory(categoryService.findById(categoryId));
             }
             transactionService.save(transaction);
         }
@@ -218,39 +210,42 @@ public class TransactionController {
 
     @ResponseBody
     @GetMapping("/delete/{id}")
-    public String delete(@PathVariable Long id, Model model) {
-        Project project1 = projectService.findById(2l);
+    public String delete(@PathVariable Long id, @AuthenticationPrincipal CurrentUser currentUser) {
+        Project project = projectService.findCurrentByUser(currentUser.getUser());
         Transaction transaction = transactionService.findById(id);
-        transaction.categories.clear();
-        transaction.pendingCategories.clear();
-        transaction.rejectedCategories.clear();
-        transactionService.delete(transaction);
+        if (transaction.project.equals(project)) {
+            transaction.categories.clear();
+            transaction.pendingCategories.clear();
+            transaction.rejectedCategories.clear();
+            transactionService.delete(transaction);
+        }
         return "";
     }
+
     @GetMapping("/assign")
-    public String assign( Model model) {
-        Project project2 = projectService.findById(2l);
-        transactionService.assignDefaultCategoriesInTransactions(project2);
-        List<Transaction> transactionsList = transactionService.findByProjectId(2l);
+    public String assign(Model model, @AuthenticationPrincipal CurrentUser currentUser) {
+        Project project = currentUser.getUser().getCurrentProject();
+        transactionService.assignDefaultCategoriesInTransactions(project);
+        List<Transaction> transactionsList = transactionService.findByProjectId(project.getId());
         model.addAttribute("tl", transactionsList);
         return "redirect:/transaction/list";
     }
 
     @GetMapping("/table/{year}/{month}")
-    public String table( Model model, @PathVariable String year, @PathVariable String month) {
-        Project project2 = projectService.findById(2l);
-        List<Transaction> transactionsList = transactionService.transactionsByDate(year, month, project2);
-        List<Category> categories = categoryService.findByProjectId(2l);
+    public String table(Model model, @PathVariable String year, @PathVariable String month, @AuthenticationPrincipal CurrentUser currentUser) {
+        Project project = currentUser.getUser().getCurrentProject();
+        List<Transaction> transactionsList = transactionService.transactionsByDate(year, month, project);
+        List<Category> categories = categoryService.findByProjectId(project.getId());
         model.addAttribute("categoriesList", categories);
         model.addAttribute("tl", transactionsList);
         return "transactions/transaction-table-rows";
     }
 
     @GetMapping("/table/gettransaction/{transactionId}")
-    public String tableRow( Model model, @PathVariable Long transactionId) {
-        Project project2 = projectService.findById(2l);
-        List <Transaction> transactions = Arrays.asList(transactionService.findById(transactionId));
-        List<Category> categories = categoryService.findByProjectId(2l);
+    public String tableRow(Model model, @PathVariable Long transactionId, @AuthenticationPrincipal CurrentUser currentUser) {
+        Project project = currentUser.getUser().getCurrentProject();
+        List<Transaction> transactions = Arrays.asList(transactionService.findById(transactionId));
+        List<Category> categories = categoryService.findByProjectId(project.getId());
         model.addAttribute("categoriesList", categories);
         model.addAttribute("tl", transactions);
         return "transactions/transaction-table-rows";
